Revert naming changes and formatting
