+++ core/android/src/main/kotlin/com/walletconnect/android/CoreProtocol.kt
@@ -42,78 +42,78 @@ import org.koin.core.qualifier.named
 import org.koin.dsl.module
 
 class CoreProtocol(private val koinApp: KoinApplication = wcKoinApp) : CoreInterface {
-	override val Pairing: PairingInterface = PairingProtocol(koinApp)
-	override val PairingController: PairingControllerInterface = PairingController(koinApp)
-	override var Relay = RelayClient(koinApp)
+    override val Pairing: PairingInterface = PairingProtocol(koinApp)
+    override val PairingController: PairingControllerInterface = PairingController(koinApp)
+    override var Relay = RelayClient(koinApp)
 
-	@Deprecated(message = "Replaced with Push")
-	override val Echo: PushInterface = PushClient
-	override val Push: PushInterface = PushClient
-	override val Verify: VerifyInterface = VerifyClient(koinApp)
-	override val Explorer: ExplorerInterface = ExplorerProtocol(koinApp)
+    @Deprecated(message = "Replaced with Push")
+    override val Echo: PushInterface = PushClient
+    override val Push: PushInterface = PushClient
+    override val Verify: VerifyInterface = VerifyClient(koinApp)
+    override val Explorer: ExplorerInterface = ExplorerProtocol(koinApp)
 
-	init {
-		plantTimber()
-	}
+    init {
+        plantTimber()
+    }
 
-	override fun setDelegate(delegate: CoreInterface.Delegate) {
-		Pairing.setDelegate(delegate)
-	}
+    override fun setDelegate(delegate: CoreInterface.Delegate) {
+        Pairing.setDelegate(delegate)
+    }
 
-	companion object {
-		val instance = CoreProtocol()
-	}
+    companion object {
+        val instance = CoreProtocol()
+    }
 
-	override fun initialize(
-		metaData: Core.Model.AppMetaData,
-		relayServerUrl: String,
-		connectionType: ConnectionType,
-		application: Application,
-		relay: RelayConnectionInterface?,
-		keyServerUrl: String?,
-		networkClientTimeout: NetworkClientTimeout?,
-		telemetryEnabled: Boolean,
-		onError: (Core.Model.Error) -> Unit
-	) {
-		try {
-			val bundleId: String = application.packageName
-			with(koinApp) {
-				androidContext(application)
-				require(relayServerUrl.isValidRelayServerUrl()) { "Check the schema and projectId parameter of the Server Url" }
-				modules(
-					module { single { ProjectId(relayServerUrl.projectId()) } },
-					module { single(named(AndroidCommonDITags.TELEMETRY_ENABLED)) { TelemetryEnabled(telemetryEnabled) } },
-					coreAndroidNetworkModule(relayServerUrl, connectionType.toCommonConnectionType(), BuildConfig.SDK_VERSION, networkClientTimeout, bundleId),
-					coreCommonModule(),
-					coreCryptoModule(),
-				)
+    override fun initialize(
+        metaData: Core.Model.AppMetaData,
+        relayServerUrl: String,
+        connectionType: ConnectionType,
+        application: Application,
+        relay: RelayConnectionInterface?,
+        keyServerUrl: String?,
+        networkClientTimeout: NetworkClientTimeout?,
+        telemetryEnabled: Boolean,
+        onError: (Core.Model.Error) -> Unit
+    ) {
+        try {
+            val bundleId: String = application.packageName
+            with(koinApp) {
+                androidContext(application)
+                require(relayServerUrl.isValidRelayServerUrl()) { "Check the schema and projectId parameter of the Server Url" }
+                modules(
+                    module { single { ProjectId(relayServerUrl.projectId()) } },
+                    module { single(named(AndroidCommonDITags.TELEMETRY_ENABLED)) { TelemetryEnabled(telemetryEnabled) } },
+                    coreAndroidNetworkModule(relayServerUrl, connectionType.toCommonConnectionType(), BuildConfig.SDK_VERSION, networkClientTimeout, bundleId),
+                    coreCommonModule(),
+                    coreCryptoModule(),
+                )
 
-				if (relay == null) {
-					Relay.initialize { error -> onError(Core.Model.Error(error)) }
-				}
+                if (relay == null) {
+                    Relay.initialize { error -> onError(Core.Model.Error(error)) }
+                }
 
-				modules(
-					coreStorageModule(bundleId = bundleId),
-					pushModule(),
-					module { single { relay ?: Relay } },
-					module { single { with(metaData) { AppMetaData(name = name, description = description, url = url, icons = icons, redirect = Redirect(redirect)) } } },
-					module { single { Echo } },
-					module { single { Push } },
-					module { single { Verify } },
-					coreJsonRpcModule(),
-					corePairingModule(Pairing, PairingController, bundleId),
-					keyServerModule(keyServerUrl),
-					explorerModule(),
-					web3ModalModule(),
-					pulseModule(bundleId)
-				)
-			}
+                modules(
+                    coreStorageModule(bundleId = bundleId),
+                    pushModule(),
+                    module { single { relay ?: Relay } },
+                    module { single { with(metaData) { AppMetaData(name = name, description = description, url = url, icons = icons, redirect = Redirect(redirect)) } } },
+                    module { single { Echo } },
+                    module { single { Push } },
+                    module { single { Verify } },
+                    coreJsonRpcModule(),
+                    corePairingModule(Pairing, PairingController, bundleId),
+                    keyServerModule(keyServerUrl),
+                    explorerModule(),
+                    web3ModalModule(),
+                    pulseModule(bundleId)
+                )
+            }
 
-			Verify.initialize()
-			Pairing.initialize()
-			PairingController.initialize()
-		} catch (e: Exception) {
-			onError(Core.Model.Error(e))
-		}
-	}
+            Verify.initialize()
+            Pairing.initialize()
+            PairingController.initialize()
+        } catch (e: Exception) {
+            onError(Core.Model.Error(e))
+        }
+    }
 }
