+++ protocol/sign/src/androidTest/kotlin/com/walletconnect/sign/test/client/SessionAuthenticateInstrumentedAndroidTest.kt
@@ -25,7 +25,7 @@ import org.junit.Test
 import org.web3j.utils.Numeric
 import timber.log.Timber
 
-class ErrorAuthenticateInstrumentedAndroidTest {
+class SessionAuthenticateInstrumentedAndroidTest {
     @get:Rule
     val scenarioExtension = SignClientInstrumentedActivityScenario()
