+++ protocol/sign/src/test/kotlin/com/walletconnect/sign/engine/use_case/calls/ExtendSessionUseCaseTest.kt
@@ -7,7 +7,6 @@ import com.walletconnect.foundation.common.model.PublicKey
 import com.walletconnect.foundation.common.model.Topic
 import com.walletconnect.foundation.util.Logger
 import com.walletconnect.sign.common.exceptions.NotSettledSessionException
-import com.walletconnect.sign.common.exceptions.UnauthorizedPeerException
 import com.walletconnect.sign.common.model.vo.sequence.SessionVO
 import com.walletconnect.sign.storage.sequence.SessionStorageRepository
 import io.mockk.every
@@ -18,7 +17,7 @@ import org.junit.Assert.fail
 import org.junit.Before
 import org.junit.Test
 
-class ExtendErrorUseCaseTest {
+class ExtendSessionUseCaseTest {
     private val jsonRpcInteractor = mockk<JsonRpcInteractorInterface>()
     private val sessionStorageRepository = mockk<SessionStorageRepository>()
     private val logger = mockk<Logger>()
