+++ protocol/sign/src/test/kotlin/com/walletconnect/sign/engine/use_case/calls/ProposeSessionUseCaseTest.kt
@@ -19,7 +19,7 @@ import org.junit.Assert.fail
 import org.junit.Before
 import org.junit.Test
 
-class ProposeErrorUseCaseTest {
+class ProposeSessionUseCaseTest {
     private val jsonRpcInteractor = mockk<JsonRpcInteractorInterface>()
     private val crypto = mockk<KeyManagementRepository>()
     private val proposalStorageRepository = mockk<ProposalStorageRepository>()
